+++ src/components/beers-filter/BeerListFilterBottomSheet/BeerCountryFilterTab/index.ts
@@ -1,0 +1,1 @@
+export { default } from './BeerCountryFilterTab';
